docker_container/requirements.txt
anyio>=3.7.0
click>=8.1.3
exceptiongroup>=1.1.2
fastapi>=0.99.1
h11>=0.14.0
idna>=3.4
Pillow>=10.0.0
pip>=22.0.2
pydantic>=1.10.11
setuptools>=59.6.0
sniffio>=1.3.0
starlette>=0.27.0
treepoem>=3.19.0
typing_extensions>=4.7.1
uvicorn>=0.22.0
